Implement Amazon and BC4 products as subclasses
