+++ feature/amazon/src/main/java/com/revenuecat/purchases/amazon/storeProductConversions.kt
@@ -1,26 +1,47 @@
 package com.revenuecat.purchases.amazon
 
-import com.amazon.device.iap.internal.model.ProductBuilder
+import android.os.Parcelable
 import com.amazon.device.iap.model.Product
+import com.revenuecat.purchases.ProductType
 import com.revenuecat.purchases.common.LogIntent
 import com.revenuecat.purchases.common.MICROS_MULTIPLIER
 import com.revenuecat.purchases.common.log
 import com.revenuecat.purchases.models.StoreProduct
-import com.revenuecat.purchases.models.StoreProductImpl
+import com.revenuecat.purchases.models.ComparableData
+import com.revenuecat.purchases.parceler.JSONObjectParceler
+import kotlinx.parcelize.Parcelize
+import kotlinx.parcelize.TypeParceler
 import org.json.JSONObject
 import java.math.BigDecimal
 import java.util.regex.Pattern
-import com.amazon.device.iap.model.ProductType as AmazonProductType
 
-val StoreProduct.amazonProduct: Product
-    get() = ProductBuilder()
-        .setSku(originalJson.getString("sku"))
-        .setProductType(originalJson.getProductType("productType"))
-        .setDescription(originalJson.getString("description"))
-        .setPrice(originalJson.getString("price"))
-        .setSmallIconUrl(originalJson.getString("smallIconUrl"))
-        .setTitle(originalJson.getString("title"))
-        .setCoinsRewardAmount(originalJson.getInt("coinsRewardAmount")).build()
+@Parcelize
+@TypeParceler<JSONObject, JSONObjectParceler>()
+data class AmazonStoreProduct(
+    override val sku: String,
+    override val type: ProductType,
+    override val price: String,
+    override val priceAmountMicros: Long,
+    override val priceCurrencyCode: String,
+    override val originalPrice: String?,
+    override val originalPriceAmountMicros: Long,
+    override val title: String,
+    override val description: String,
+    override val subscriptionPeriod: String?,
+    override val freeTrialPeriod: String?,
+    override val introductoryPrice: String?,
+    override val introductoryPriceAmountMicros: Long,
+    override val introductoryPricePeriod: String?,
+    override val introductoryPriceCycles: Int,
+    override val iconUrl: String,
+    override val originalJson: JSONObject,
+    val amazonProduct: Product
+) : StoreProduct, Parcelable {
+
+    // We use this to not include the originalJSON in the equals
+    override fun equals(other: Any?) = other is StoreProduct && ComparableData(this) == ComparableData(other)
+    override fun hashCode() = ComparableData(this).hashCode()
+}
 
 fun Product.toStoreProduct(marketplace: String): StoreProduct? {
     if (price == null) {
@@ -31,7 +52,7 @@ fun Product.toStoreProduct(marketplace: String): StoreProduct? {
     // the local currency of each marketplace where they can be sold, and customers will see IAP items in English.
     val (currencyCode, priceAmountMicros) = price.extractPrice(marketplace)
 
-    return StoreProductImpl(
+    return AmazonStoreProduct(
         sku,
         productType.toRevenueCatProductType(),
         price,
@@ -48,7 +69,8 @@ fun Product.toStoreProduct(marketplace: String): StoreProduct? {
         introductoryPricePeriod = null,
         introductoryPriceCycles = 0,
         iconUrl = smallIconUrl,
-        originalJson = toJSON()
+        originalJson = toJSON(),
+        amazonProduct = this
     )
 }
 
@@ -102,6 +124,3 @@ internal fun String.parsePriceUsingRegex(): BigDecimal? {
         BigDecimal(price)
     }
 }
-
-private fun JSONObject.getProductType(productType: String) =
-    AmazonProductType.values().firstOrNull { it.name == this.getString(productType) }
