+++ feature/google/src/main/java/com/revenuecat/purchases/google/storeProductConversions.kt
@@ -1,11 +1,10 @@
 package com.revenuecat.purchases.google
 
 import com.android.billingclient.api.SkuDetails
-import com.revenuecat.purchases.models.StoreProductImpl
 import org.json.JSONObject
 
 fun SkuDetails.toStoreProduct() =
-    StoreProductImpl(
+    BC4StoreProduct(
         sku,
         type.toRevenueCatProductType(),
         price,
