--- public/src/main/java/com/revenuecat/purchases/models/productDetailsExtensions.kt
@@ -1,11 +1,0 @@
-@file:JvmName("StoreProductHelpers")
-
-package com.revenuecat.purchases.models
-
-import com.android.billingclient.api.SkuDetails
-
-/**
- * Returns the original SkuDetails that was used to build the StoreProduct object.
- */
-val StoreProduct.skuDetails: SkuDetails
-    get() = SkuDetails(this.originalJson.toString())
