+++ public/src/main/java/com/revenuecat/purchases/parceler/JSONObjectParceler.kt
@@ -5,7 +5,7 @@ import kotlinx.parcelize.Parceler
 import org.json.JSONObject
 
 /** @suppress */
-internal object JSONObjectParceler : Parceler<JSONObject> {
+object JSONObjectParceler : Parceler<JSONObject> {
 
     override fun create(parcel: Parcel): JSONObject {
         return JSONObject(parcel.readString())
